Rename benchmark project, extract model types, and use FrozenDictionary
- Rename BenchmarkSuite1 to MYA.RequestProtect.Benchmarks with multi-TFM
  support (net8.0/net9.0) and direct project reference
- Extract WhitelistEntry and HeaderEntry structs into Models/ folder
- Replace ConcurrentDictionary regex cache with pre-built FrozenDictionary
- Use List<T> with CollectionsMarshal.AsSpan for parsed whitelist/headers
- Update CLAUDE.md to reflect benchmark project rename
