+++ src/MYA.RequestProtect.Benchmarks/MYA.RequestProtect.Benchmarks.csproj
@@ -1,8 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFrameworks>net8.0;net9.0</TargetFrameworks>
     <OutputType>Exe</OutputType>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
@@ -11,6 +13,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MYA.RequestProtect.TestSite\MYA.RequestProtect.TestSite.csproj" />
+    <ProjectReference Include="..\MYA.RequestProtect\MYA.RequestProtect.csproj" />
   </ItemGroup>
 </Project>
